pyproject.toml
[tool.poetry]
name = "echoes-of-eden"
version = "0.1.0"
description = ""
authors = ["Nate Stott <nate.stott@pm.me>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
matplotlib = "^3.9.2"
pathfinding = "^1.0.11"
python-dotenv = "^1.0.1"
loguru = "^0.7.2"
seaborn = "^0.13.2"

[tool.poetry.group.dev.dependencies]
pylint = "^3.3.1"
mypy = "^1.13.0"
black = "^24.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
